A first try to account for Surveyor's dual tagging mechanisms.
Surveyor permits questions, answers, labels, and some similar constructs
to have two sources for the reference identifier.  The first source is
via the method called in the DSL (e.g. q_foo creates a question with
reference identifier "foo"); the second source is via the
:reference_identifier option.

The interplay of these two sources is not documented; this is just my
first real guess.