Differentiate conditions for validations and questions.
Some conditions are associated with validations, and therefore belong to
answers.  Other conditions are associated with questions, and therefore
belong to questions.  (They may also reference other questions and their
answers.)

I previously did not differentiate between these two types of
conditions, because it was not useful to do so.  However, now that I'm
looking into building transitive closures of survey dependencies, I have
a need to be able to resolve references in conditions, and (so far)
being able to discriminate between these two types of conditions seems
useful.  We'll see if that stays the case.